Add currency stas overview Closes https://github.com/invenhost/InvenTree/issues/115
